server/requirements.txt
Django==1.9
djangorestframework==3.3.2
django-filter==0.11.0
django-oauth-toolkit==0.10.0
